Testing pull/push with github

+++ src/com/lvl6/server/GameServer.java
@@ -59,7 +59,7 @@ public class GameServer extends Thread{
   private int portNum;
   
   //current client version to see if it is still playable
-  public static float clientVersionNumber = (float) 1.0;
+  public static float clientVersionNumber = 1.0f;
 
   public static void main(String args[]) {
     if (args.length == 2) {
